Add logic for finances and statistics feature

+++ src/main/java/com/example/poc/masterspreparation/dto/AttendanceScheduleDto.java
@@ -17,5 +17,5 @@ public class AttendanceScheduleDto {
 
     private String workerEmail;
 
-    private Double sum;
+    private Integer sum;
 }
